+++ src/Compiler/IL/TempLocalAllocator.cs
@@ -507,6 +507,8 @@ internal static partial class LIRInstructionInfo
                 visitor.Visit(value.Receiver); visitor.Visit(value.A0); visitor.Visit(value.A1); visitor.Visit(value.A2); visitor.Visit(value.A3); visitor.Visit(value.A4); break;
             case LIRCallGuardedStringIntrinsic value:
                 visitor.Visit(value.Receiver); VisitList(value.Arguments, ref visitor); break;
+            case LIRCallGuardedIntrinsicMember value:
+                visitor.Visit(value.Receiver); VisitList(value.Arguments, ref visitor); break;
             case LIRCallComputedMemberFixed value:
                 visitor.Visit(value.Receiver); visitor.Visit(value.PropertyKey); VisitList(value.Arguments, ref visitor); break;
             case LIRCallNodeModuleContractMember value:
@@ -1012,6 +1014,9 @@ internal static partial class LIRInstructionInfo
                 return true;
             case LIRCallGuardedStringIntrinsic guardedString:
                 defined = guardedString.Result;
+                return true;
+            case LIRCallGuardedIntrinsicMember guardedIntrinsic:
+                defined = guardedIntrinsic.Result;
                 return true;
             case LIRCallComputedMemberFixed callComputedMember:
                 defined = callComputedMember.Result;