+++ tests/Jroc.Tests/Object/JavaScript/ObjectLiteral_Method_DynamicOverride.js
@@ -1,0 +1,13 @@
+"use strict";
+
+let receiver = {
+    value() {
+        return "original";
+    }
+};
+
+console.log(receiver.value());
+receiver.value = function () {
+    return "override";
+};
+console.log(receiver.value());